:sparkles: feat: Implement UserController with user management endpoints
- Added UserController for handling user-related operations including:
  - Debugging user-team associations
  - Retrieving current user information
  - Adding teams to users
- Introduced AuthRequest DTO for authentication requests.
- Created CreateEquipeRequest and CreateTaskRequest DTOs for future use.
- Added LoginRequest and RegisterRequest DTOs for user login and registration.
- Implemented AuthResponse DTO for authentication responses.
- Defined EquipeResponse, ErrorResponse, TaskResponse, and UserResponse DTOs for structured responses.
- Developed mappers for Equipe, Task, and User entities.
- Created embedded models for task responsibilities and task information.
- Established entity models for Author, Comment, Equipe, Membro, Responsible, Task, and User.
- Implemented repositories for Equipe, Task, and User entities.
- Developed AuthService, EquipeService, and TaskService for business logic handling.
- Added task management functionalities including task creation, update, and deletion.
