+++ back-end/api-3/src/main/java/com/api_3/api_3/dto/response/TaskResponse.java
@@ -1,0 +1,5 @@
+package com.api_3.api_3.dto.response;
+
+public class TaskResponse {
+    
+}
