+++ back-end/api-3/src/main/java/com/api_3/api_3/mapper/TaskMapper.java
@@ -1,0 +1,5 @@
+package com.api_3.api_3.mapper;
+
+public class TaskMapper {
+    
+}
