+++ back-end/api-3/src/main/java/com/api_3/api_3/model/embedded/ResponsavelTask.java
@@ -1,0 +1,12 @@
+package com.api_3.api_3.model.embedded;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class ResponsavelTask {
+    private String uuid;
+    private String name;
+    private String img;
+}
