+++ back-end/api-3/src/main/java/com/api_3/api_3/model/embedded/TaskInfo.java
@@ -1,0 +1,16 @@
+package com.api_3.api_3.model.embedded;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class TaskInfo {
+    private String uuid;
+    private String title;
+    private String due_date;
+    private String status;
+    private String prioridade;
+    private String equipe_uuid;
+    private ResponsavelTask responsavel;
+}
