+++ back-end/api-3/src/main/java/com/api_3/api_3/model/entity/Author.java
@@ -1,0 +1,12 @@
+package com.api_3.api_3.model.entity;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class Author {
+    private String uuid;
+    private String name;
+    private String url_img;
+}
