+++ back-end/api-3/src/main/java/com/api_3/api_3/model/entity/Comment.java
@@ -1,0 +1,14 @@
+package com.api_3.api_3.model.entity;
+
+import java.util.Date;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class Comment {
+    private String content;
+    private Date created_at;
+    private Author author;
+}
