+++ back-end/api-3/src/main/java/com/api_3/api_3/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.api_3.api_3.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+import com.api_3.api_3.model.entity.User;
+
+@Repository
+public interface UserRepository extends MongoRepository<User, String>{
+    Optional<User> findByEmail(String email);    
+}
