--- back-end/api-3/src/main/java/com/api_3/api_3/task/model/Comment.java
@@ -1,13 +1,0 @@
-package com.api_3.api_3.task.model;
-
-import java.util.Date;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class Comment {
-    private String content;
-    private Date created_at;
-    private Author author;
-}
