--- back-end/api-3/src/main/java/com/api_3/api_3/task/model/Responsible.java
@@ -1,12 +1,0 @@
-package com.api_3.api_3.task.model;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class Responsible {
-    private String uuid;
-    private String name;
-    private String url_img;
-}
