--- back-end/api-3/src/main/java/com/api_3/api_3/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.api_3.api_3.user.repository;
-
-import java.util.Optional;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-import com.api_3.api_3.user.model.User;
-
-@Repository
-public interface UserRepository extends MongoRepository<User, String>{
-    Optional<User> findByEmail(String email);    
-}
